Moved dfly and Dual to the right files, moved PwMap up in the include list
