+++ src/InvariantMeasures.jl
@@ -17,19 +17,20 @@ include("BasisDefinition.jl")
 include("GenericAssembler.jl")
 include("GenericEstimate.jl")
 
+include("PwDynamicDefinition.jl")
 include("Mod1Dynamic.jl")
 
 include("UlamBasis.jl")
 include("HatBasis.jl")
 
-using .DynamicDefinition, .BasisDefinition, .Mod1DynamicDefinition, .Contractors
+using .DynamicDefinition, .BasisDefinition, .Mod1DynamicDefinition, .Contractors, .PwDynamicDefinition
 
 include("Norms.jl")
 include("pitrig.jl")
 include("NormsOfPowers.jl")
 
 export NormKind, L1, Linf, Lipschitz, TotalVariation
-export Ulam, Mod1Dynamic, Basis, Dynamic, assemble, preim, Hat, EquispacedPartition, norms_of_powers, sinpi, cospi, dfly, DiscretizedOperator, opnormbound, weak_norm, strong_norm, aux_norm, integral_covector, distance_from_invariant
+export Ulam, PwMap, Mod1Dynamic, Basis, Dynamic, assemble, preim, Hat, EquispacedPartition, norms_of_powers, sinpi, cospi, dfly, DiscretizedOperator, opnormbound, weak_norm, strong_norm, aux_norm, integral_covector, distance_from_invariant
 import ValidatedNumerics: Interval
 export Interval
 
@@ -37,10 +38,5 @@ export Interval
 include("InducedLSV.jl")
 using .InducedLSVMapDefinition
 export ApproxInducedLSV
-
-include("PwDynamicDefinition.jl")
-using .PwDynamicDefinition
-export PwMap
-
 
 end
